seeding tables with initial datas

+++ Taxation.API/Models/MonthlyTaxRequest.cs
@@ -3,7 +3,7 @@
     public class MonthlyTaxRequest
     {
         public int Municipality { get; set; }
-        public decimal Tax { get; set; }
+        public float Tax { get; set; }
         public DateTimeOffset StartDate { get; set; }
         public DateTimeOffset EndDate { get; set; }
     }
